+++ crates/assistant/Cargo.toml
@@ -33,7 +33,7 @@ client.workspace = true
 clock.workspace = true
 collections.workspace = true
 command_palette_hooks.workspace = true
-context_servers.workspace = true
+context_server.workspace = true
 db.workspace = true
 editor.workspace = true
 feature_flags.workspace = true
