--- crates/assistant/src/tools.rs
@@ -1,2 +1,0 @@
-pub mod context_server_tool;
-pub mod now_tool;
