+++ crates/assistant_tools/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "assistant_tools"
+version = "0.1.0"
+edition = "2021"
+publish = false
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/assistant_tools.rs"
+
+[dependencies]
+anyhow.workspace = true
+assistant_tool.workspace = true
+chrono.workspace = true
+gpui.workspace = true
+schemars.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+workspace.workspace = true
