+++ crates/assistant_tools/LICENSE-GPL
