+++ crates/assistant_tools/src/now_tool.rs
