+++ crates/context_server/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "context_servers"
+name = "context_server"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -9,12 +9,14 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/context_servers.rs"
+path = "src/context_server.rs"
 
 [dependencies]
 anyhow.workspace = true
+assistant_tool.workspace = true
 collections.workspace = true
 command_palette_hooks.workspace = true
+context_server_settings.workspace = true
 extension.workspace = true
 futures.workspace = true
 gpui.workspace = true
@@ -22,10 +24,11 @@ log.workspace = true
 parking_lot.workspace = true
 postage.workspace = true
 project.workspace = true
-schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 settings.workspace = true
 smol.workspace = true
+ui.workspace = true
 url = { workspace = true, features = ["serde"] }
 util.workspace = true
+workspace.workspace = true
