+++ crates/context_server/LICENSE-GPL
@@ -1,0 +1,1 @@
+../../LICENSE-GPL
