+++ crates/context_server/src/context_server.rs
@@ -1,4 +1,5 @@
 pub mod client;
+mod context_server_tool;
 mod extension_context_server;
 pub mod manager;
 pub mod protocol;
@@ -6,10 +7,10 @@ mod registry;
 pub mod types;
 
 use command_palette_hooks::CommandPaletteFilter;
+pub use context_server_settings::{ContextServerSettings, ServerCommand, ServerConfig};
 use gpui::{actions, AppContext};
-use settings::Settings;
 
-use crate::manager::ContextServerSettings;
+pub use crate::context_server_tool::ContextServerTool;
 pub use crate::registry::ContextServerFactoryRegistry;
 
 actions!(context_servers, [Restart]);
@@ -18,7 +19,7 @@ actions!(context_servers, [Restart]);
 pub const CONTEXT_SERVERS_NAMESPACE: &'static str = "context_servers";
 
 pub fn init(cx: &mut AppContext) {
-    ContextServerSettings::register(cx);
+    context_server_settings::init(cx);
     ContextServerFactoryRegistry::default_global(cx);
     extension_context_server::init(cx);
