+++ crates/zed/Cargo.toml
@@ -20,6 +20,7 @@ anyhow.workspace = true
 assets.workspace = true
 assistant.workspace = true
 assistant2.workspace = true
+assistant_tools.workspace = true
 async-watch.workspace = true
 audio.workspace = true
 auto_update.workspace = true
